api_test/requirement.txt
pip==20.0.2
Django==3.0.3
django-filter==2.2.0
django-polymorphic==2.1.2
djangorestframework==3.11.0
djangorestframework-csv==2.1.0
djangorestframework-jsonapi==3.1.0
drf-yasg==1.17.1